requirements.txt
langchain
openai
python-dotenv
streamlit
langchain_community
langchain_openai
requests
tweepy
selenium
